JAMES-4135 Make CriterionConverter into an interface
